updating rampage motif usage

+++ training_utils_rampage_finetune.py
@@ -313,7 +313,7 @@ def deserialize_tr(serialized_example, g, use_motif_activity,
                                   axis=1)
     peaks_c_crop = tf.slice(peaks_center, [crop_size,0], [output_length-2*crop_size,-1]) # crop at the outset 
     # TF activity, cast to float32 and expand dims to allow for processing by model input FC layers
-    motif_activity = tf.ensure_shape(tf.io.parse_tensor(data['motif_activity'], out_type=tf.float16), [693])
+    motif_activity = tf.ensure_shape(tf.io.parse_tensor(data['motif_activity'], out_type=tf.float32), [693])
     motif_activity = tf.cast(motif_activity,dtype=tf.float32)
     motif_activity = tf.expand_dims(motif_activity,axis=0)
     motif_activity = (motif_activity - motif_means) / (motif_std + 1.0e-06)
@@ -500,7 +500,7 @@ def deserialize_val(serialized_example, g_val, use_motif_activity,
     sequence = one_hot(tf.strings.substr(data['sequence'], shift,input_length))
 
     # motif activity, cast to float32 and expand dims to allow for processing by model input FC layers
-    motif_activity = tf.ensure_shape(tf.io.parse_tensor(data['motif_activity'], out_type=tf.float16), [693])
+    motif_activity = tf.ensure_shape(tf.io.parse_tensor(data['motif_activity'], out_type=tf.float32), [693])
     motif_activity = tf.cast(motif_activity,dtype=tf.float32)
     motif_activity = tf.expand_dims(motif_activity,axis=0)
     motif_activity = (motif_activity - motif_means) / (motif_std + 1.0e-06)
